kernel: Introduce SBalance IRQ balancer (port Linux 4.4.y)
This is a simple IRQ balancer that polls every X number of milliseconds and
moves IRQs from the most interrupt-heavy CPU to the least interrupt-heavy
CPUs until the heaviest CPU is no longer the heaviest. IRQs are only moved
from one source CPU to any number of destination CPUs per balance run.
Balancing is skipped if the gap between the most interrupt-heavy CPU and
the least interrupt-heavy CPU is below the configured threshold of
interrupts.

The heaviest IRQs are targeted for migration in order to reduce the number
of IRQs to migrate. If moving an IRQ would reduce overall balance, then it
won't be migrated.

The most interrupt-heavy CPU is calculated by scaling the number of new
interrupts on that CPU to the CPU's current capacity. This way, interrupt
heaviness takes into account factors such as thermal pressure and time
spent processing interrupts rather than just the sheer number of them. This
also makes SBalance aware of CPU asymmetry, where different CPUs can have
different performance capacities and be proportionally balanced.

Changes for Linux 4.4:
  Added <linux/kthread.h> which is later called by ../sched/sched.h
  Delete the const pointer of bal_irq_move_node_cmp
  refer to ("treewide: Change list_sort to use const pointers")

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I60544cdc9eb8ac751ccac22d23314c4e668e3cef